Refactor variable names and data types across classes
Changed the variable name from 'key' to 'search' in valueList method across the engine classes to be more descriptive about its functionality. Moreover, adjusted the data types for variables in setExpiration methods from float to integer to ensure consistency throughout the code. An enhancement in Periods class has been made to make sure the difference in seconds is always an absolute integer.

+++ src/DataEngines/DataEngine.php
@@ -27,5 +27,5 @@ Interface DataEngine
      */
     public function timeLeft(string $key): int;
 
-    public function setExpiration(string $key, float $time): bool;
+    public function setExpiration(string $key, int $time): bool;
 }
